prevent inline static final null literal

+++ cpp/include/lyra/types.h
@@ -1,17 +1,18 @@
 #ifndef LYRA_CPP_TYPES
 #define LYRA_CPP_TYPES
 
+#include <stdint.h>
 #include <jni.h>
 
 namespace lyra {
 	__attribute__((always_inline)) inline void* address(jlong addr)
 	{
-		return (void*)(uintptr_t)addr;
+		return (void*)(uint64_t)addr;
 	}
 
 	__attribute__((always_inline)) inline jlong address(void* ptr)
 	{
-		return (uintptr_t)ptr;
+		return (uint64_t)ptr;
 	}
 }
